arpspoofing-to-irc: Capture ARP packets only.
Capturing on all types of packets only increases the attack vector, as
attackers might be able to inject arbitrary strings into other packets
and thus trick tcpdump into printing it as-is to the output pipe.

Apart from that, depending on the load of the machine, the kernel might
drop packets if the capture buffer runs out of space. And if an attacker
manages to overload the machine it's also possible to bypass the ARP
spoofing detection mechanism.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>