check_arp_replies: Improve tcpdump line checking.
The checks on the output of tcpdump previously was pretty lax and
offered enough room for injecting arbitrary strings. Also, doing a match
with a regular expression and then processing the line again with a even
less strict approach is not only very error-prone but also redundant and
can be done in one go.

As for the new regular expression, I'm allowing . and : in IPs in the
event that the IP should be an IPv6 address. I'm not sure whether it
really has the same tcpdump output as the IPv4 address, but it doesn't
hurt in terms of security, because : is not a character that should be
able to spoil the fun.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>