check_arp_replies: Limit tcpdump match lengths.
This limits the match groups of the regular expression so that the worst
that could happen would be a MAC address with 4 hexdigits instead of 2,
but no such thing as flooding the IRC channel with a shitload of any
combination out of hexdigits and separators.

I'm using XX and replacing it in the regular expression to avoid doing
[0-9a-fA-F]{count} over and over again.

Also, in order to be more strict in the matching process, I've added a
comma at the end of the regular expression to make sure we get a match
of only valid (to some extent, as noted before) MAC addresses and not
miss things such as ab:cd:ef:01:23:45:67, which clearly should be
invalid and thus be dropped.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>